Rename to ConfidenceSlider and add configurable step size and modes
