+++ demo/multi-class.html
@@ -127,8 +127,8 @@
                     AllowedToolboxItem.AnnotationID, 
                     AllowedToolboxItem.ClassCounter,
                     AllowedToolboxItem.AnnotationResize, 
-                    AllowedToolboxItem.KeypointSlider,  
-                    AllowedToolboxItem.ConfidenceFilter,
+                    AllowedToolboxItem.KeypointSlider,
+                    AllowedToolboxItem.ConfidenceSlider,
                     AllowedToolboxItem.FilterDistance,
                     AllowedToolboxItem.RecolorActive,
                 ],
