+++ index.d.ts
@@ -36,13 +36,13 @@ export type Offset = {
 /**
  * Valid keys for the DeprecatedBy type
  */
-export type ValidDeprecatedBy = "human" | "confidence_filter" | "distance_from_row" | "spatial_confidence_filter";
+export type ValidDeprecatedBy = "human" | "confidence_filter" | "distance_from_row" | "confidence_slider";
 
 export type DeprecatedBy = {
     human?: boolean;
     confidence_filter?: boolean;
     distance_from_row?: boolean;
-    spatial_confidence_filter?: boolean;
+    confidence_slider?: boolean;
 };
 
 /**
@@ -124,26 +124,33 @@ export type ConfidenceThreshold = {
 };
 
 /**
- * Stores the current confidence filter threshold values.
+ * Stores the current confidence slider threshold values.
  * The key is the class id. "all" is a special key that stores the threshold applied to
  * all annotations in single-class mode.
  */
-export type ConfidenceFilterClasses = {
+export type ConfidenceSliderClasses = {
     all: ConfidenceThreshold;
     [key: string]: ConfidenceThreshold;
 };
 
 /**
- * Config object for the ConfidenceFilterItem ToolboxItem.
+ * Controls the multi-class mode of the ConfidenceSlider ToolboxItem:
+ * - "toggle": a checkbox lets the user switch between single- and multi-class filtering
+ * - "single-only": only a single global slider is shown
+ * - "multi-only": only per-class sliders are shown
  */
-export type ConfidenceFilterConfig = {
+export type ConfidenceSliderMultiClassMode = "toggle" | "single-only" | "multi-only";
+
+/**
+ * Config object for the ConfidenceSlider ToolboxItem.
+ */
+export type ConfidenceSliderConfig = {
     name?: string;
     filter_min?: number;
     filter_max?: number;
-    default_values?: ConfidenceFilterClasses;
+    default_values?: ConfidenceSliderClasses;
     step_value?: number;
-    multi_class_mode?: boolean;
-    disable_multi_class_mode?: boolean;
+    multi_class_mode?: ConfidenceSliderMultiClassMode;
     filter_on_load?: boolean;
     show_options?: boolean;
     // The spatial types to filter. Defaults to all confidence-filterable spatial types.
@@ -405,7 +412,7 @@ export class ULabel {
     ): void;
     public get_keypoint_slider_value(): number | null;
     public get_distance_filter_value(): DistanceFromPolylineClasses | null;
-    public get_confidence_filter_value(): ConfidenceFilterClasses | null;
+    public get_confidence_slider_value(): ConfidenceSliderClasses | null;
     public fly_to_next_annotation(increment: number, max_zoom?: number): boolean;
     public fly_to_annotation_id(annotation_id: string, subtask_key?: string | null, max_zoom?: number): boolean;
     public fly_to_annotation(annotation: ULabelAnnotation, subtask_key?: string, max_zoom?: number): boolean;
