+++ tests/types/index.test-d.ts
@@ -86,5 +86,5 @@ const all_items: AllowedToolboxItem[] = [
     AllowedToolboxItem.ImageFilters,
     AllowedToolboxItem.AnnotationList,
     AllowedToolboxItem.Keybinds,
-    AllowedToolboxItem.ConfidenceFilter,
+    AllowedToolboxItem.ConfidenceSlider,
 ];
